refactor: move util functions
