+++ now/utils.py
@@ -1,24 +1,18 @@
 from __future__ import annotations, print_function, unicode_literals
 
-import os
-import shutil
 import signal
 import sys
 from collections.abc import MutableMapping
 from dataclasses import dataclass
 from inspect import stack
-from typing import Any, Dict, List, Optional, TypeVar, Union
+from typing import Any, TypeVar
 
 import boto3
-import cowsay
 import docarray
 import hubble
 import yaml
 from jina.jaml import JAML
 from pyfiglet import Figlet
-
-from now.deployment.deployment import list_all_wolf, status_wolf
-from now.thirdparty.PyInquirer.prompt import prompt
 
 
 def my_handler(signum, frame, spinner):
@@ -90,38 +84,6 @@ def print_headline():
         '5GB - 8GB should be okay.'
     )
     print()
-
-
-def maybe_prompt_user(questions, attribute, **kwargs):
-    """
-    Checks the `kwargs` for the `attribute` name. If present, the value is returned directly.
-    If not, the user is prompted via the cmd-line using the `questions` argument.
-
-    :param questions: A dictionary that is passed to `PyInquirer.prompt`
-        See docs: https://github.com/CITGuru/PyInquirer#documentation
-    :param attribute: Name of the value to get. Make sure this matches the name in `kwargs`
-
-    :return: A single value of either from `kwargs` or the user cli input.
-    """
-    if kwargs and attribute in kwargs:
-        return kwargs[attribute]
-    else:
-        answer = prompt(questions)
-        return answer[attribute]
-
-
-def prompt_value(
-    name: str,
-    prompt_message: str,
-    prompt_type: str = 'input',
-    choices: Optional[List[Union[Dict, str]]] = None,
-    **kwargs: Dict,
-):
-    qs = {'name': name, 'type': prompt_type, 'message': prompt_message}
-
-    if choices is not None:
-        qs['choices'] = choices
-    return maybe_prompt_user(qs, name, **kwargs)
 
 
 def debug(msg: Any):
@@ -208,34 +170,6 @@ def get_chunk_by_field_name(doc, field_name):
         return doc.chunks[field_position]
     except Exception as e:
         print(f'An error occurred: {e}')
-
-
-def get_flow_status(action, **kwargs):
-    choices = []
-    # Add all remote Flows that exists with the namespace `nowapi`
-    alive_flows = list_all_wolf(status='Serving')
-    for flow_details in alive_flows:
-        choices.append(flow_details['name'])
-    if len(choices) == 0:
-        cowsay.cow(f'nothing to {action}')
-        return
-    else:
-        questions = [
-            {
-                'type': 'list',
-                'name': 'cluster',
-                'message': f'Which cluster do you want to {action}?',
-                'choices': choices,
-            }
-        ]
-        cluster = maybe_prompt_user(questions, 'cluster', **kwargs)
-
-    flow = [x for x in alive_flows if x['name'] == cluster][0]
-    flow_id = flow['id']
-    _result = status_wolf(flow_id)
-    if _result is None:
-        print(f'❎ Flow not found in JCloud. Likely, it has been deleted already')
-    return _result, flow_id, cluster
 
 
 # Add a custom retry exception
